LayoutTests/fast/frames/out-of-document-iframe-has-child-frame-expected.txt

CONSOLE MESSAGE: line 19: NO_MODIFICATION_ALLOWED_ERR: DOM Exception 7: An attempt was made to modify an object where modifications are not allowed.
This tests that several ways of making an iframe that isn't inserted into a document tree but has a child frame will fail.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS targetFrame1.contentWindow == undefined is true
PASS targetFrame2.contentWindow == undefined is true
PASS targetFrame3.contentWindow == undefined is true
PASS successfullyParsed is true

TEST COMPLETE
